+++ js/app/customer/customerModalCtrl.js
@@ -49,15 +49,18 @@ angular
                 // Only auto fill when add new customer and
                 // the Display Name input is not dirty
                 if (!$scope.customerForm.display_name.$dirty && !$scope.customer.id) {
-                    var currentDisplayName = '';
-                    if (!$scope.customer.given_name) {
-                        $scope.customer.given_name = '';
+                    var displayName = '';
+
+                    if ($scope.customer.family_name) {
+                        displayName = $scope.customer.family_name;
                     }
-                    if (!$scope.customer.family_name) {
-                        $scope.customer.family_name = '';
+
+                    if ($scope.customer.given_name && displayName) {
+                        displayName += ', ' + $scope.customer.given_name;
+                    } else if($scope.customer.given_name) {
+                        displayName = $scope.customer.given_name;
                     }
-                    $scope.customer.display_name = ($scope.customer.given_name +
-                        ' ' + $scope.customer.family_name).trim();
+                    $scope.customer.display_name = displayName.trim();
                 }
             };